Fix Index Bug decoding step points
