+++ libandroid-navigation/src/test/java/com/mapbox/services/android/navigation/v5/BaseTest.java
@@ -1,13 +1,19 @@
 package com.mapbox.services.android.navigation.v5;
 
-import static junit.framework.Assert.assertEquals;
-import static okhttp3.internal.Util.UTF_8;
-
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
 import com.google.gson.JsonParser;
+import com.mapbox.api.directions.v5.DirectionsAdapterFactory;
+import com.mapbox.api.directions.v5.models.DirectionsResponse;
+import com.mapbox.api.directions.v5.models.DirectionsRoute;
+import com.mapbox.services.android.navigation.v5.routeprogress.RouteProgress;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Scanner;
+
+import static junit.framework.Assert.assertEquals;
+import static okhttp3.internal.Util.UTF_8;
 
 public class BaseTest {
 
@@ -15,6 +21,7 @@ public class BaseTest {
   public static final double LARGE_DELTA = 0.1;
 
   public static final String ACCESS_TOKEN = "pk.XXX";
+  private static final String DIRECTIONS_PRECISION_6 = "directions_v5_precision_6.json";
 
   public void compareJson(String json1, String json2) {
     JsonParser parser = new JsonParser();
@@ -26,5 +33,29 @@ public class BaseTest {
     InputStream inputStream = classLoader.getResourceAsStream(filename);
     Scanner scanner = new Scanner(inputStream, UTF_8.name()).useDelimiter("\\A");
     return scanner.hasNext() ? scanner.next() : "";
+  }
+
+  protected RouteProgress buildDefaultRouteProgress() throws Exception {
+    DirectionsRoute aRoute = buildDirectionsRoute();
+    RouteProgress defaultRouteProgress = RouteProgress.builder()
+      .stepDistanceRemaining(100)
+      .legDistanceRemaining(100)
+      .distanceRemaining(100)
+      .directionsRoute(aRoute)
+      .stepIndex(0)
+      .legIndex(0)
+      .build();
+
+    return defaultRouteProgress;
+  }
+
+  protected DirectionsRoute buildDirectionsRoute() throws IOException {
+    Gson gson = new GsonBuilder()
+      .registerTypeAdapterFactory(DirectionsAdapterFactory.create()).create();
+    String body = loadJsonFixture(DIRECTIONS_PRECISION_6);
+    DirectionsResponse response = gson.fromJson(body, DirectionsResponse.class);
+    DirectionsRoute aRoute = response.routes().get(0);
+
+    return aRoute;
   }
 }
