+++ libandroid-navigation/src/test/java/com/mapbox/services/android/navigation/v5/offroute/OffRouteDetectorTest.java
@@ -3,11 +3,6 @@ package com.mapbox.services.android.navigation.v5.offroute;
 import android.location.Location;
 import android.support.annotation.NonNull;
 
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.mapbox.api.directions.v5.DirectionsAdapterFactory;
-import com.mapbox.api.directions.v5.models.DirectionsResponse;
-import com.mapbox.api.directions.v5.models.DirectionsRoute;
 import com.mapbox.api.directions.v5.models.LegStep;
 import com.mapbox.core.constants.Constants;
 import com.mapbox.geojson.LineString;
@@ -27,7 +22,6 @@ import org.mockito.MockitoAnnotations;
 import org.robolectric.RobolectricTestRunner;
 import org.robolectric.annotation.Config;
 
-import java.io.IOException;
 import java.util.List;
 
 import static junit.framework.Assert.assertFalse;
@@ -40,7 +34,6 @@ import static org.mockito.Mockito.verify;
 @Config(constants = BuildConfig.class, manifest = Config.DEFAULT_MANIFEST_NAME)
 public class OffRouteDetectorTest extends BaseTest {
 
-  private static final String DIRECTIONS_PRECISION_6 = "directions_v5_precision_6.json";
   @Mock
   private Location mockLocation;
   @Mock
@@ -263,29 +256,5 @@ public class OffRouteDetectorTest extends BaseTest {
   @NonNull
   private Point buildPointAwayFromPoint(Point point, double distanceAway, double bearing) {
     return TurfMeasurement.destination(point, distanceAway, bearing, TurfConstants.UNIT_METERS);
-  }
-
-  private RouteProgress buildDefaultRouteProgress() throws Exception {
-    DirectionsRoute aRoute = buildDirectionsRoute();
-    RouteProgress defaultRouteProgress = RouteProgress.builder()
-      .stepDistanceRemaining(100)
-      .legDistanceRemaining(100)
-      .distanceRemaining(100)
-      .directionsRoute(aRoute)
-      .stepIndex(0)
-      .legIndex(0)
-      .build();
-
-    return defaultRouteProgress;
-  }
-
-  private DirectionsRoute buildDirectionsRoute() throws IOException {
-    Gson gson = new GsonBuilder()
-      .registerTypeAdapterFactory(DirectionsAdapterFactory.create()).create();
-    String body = loadJsonFixture(DIRECTIONS_PRECISION_6);
-    DirectionsResponse response = gson.fromJson(body, DirectionsResponse.class);
-    DirectionsRoute aRoute = response.routes().get(0);
-
-    return aRoute;
   }
 }
